+++ R/register_all_transpilers.R
@@ -43,6 +43,7 @@ register_all_transpilers <- function() {
   transpilers_for_package("futurize::add-on", package = "partykit",       append_transpilers_for_partykit)
   transpilers_for_package("futurize::add-on", package = "riskRegression", append_transpilers_for_riskRegression)
   transpilers_for_package("futurize::add-on", package = "seriation",      append_transpilers_for_seriation)
+  transpilers_for_package("futurize::add-on", package = "SimDesign",      append_transpilers_for_SimDesign)
   transpilers_for_package("futurize::add-on", package = "shapr",          append_transpilers_for_shapr)
   transpilers_for_package("futurize::add-on", package = "strucchange",    append_transpilers_for_strucchange)
   transpilers_for_package("futurize::add-on", package = "tm",             append_transpilers_for_tm)